this should be in cleanup configuration now s3
